airflow-astro/requirements.txt
apache-airflow-providers-docker
apache-airflow-providers-apache-spark
apache-airflow-providers-ssh